requirements.txt
transformers==3.3.1
torch==1.6.0
numpy==1.19.1
scikit-learn==0.23.2
pyyaml==5.3.1
tqdm==4.48.2